refactor: (#11) 컴포넌트 폴더 구조 분리
