+++ src/features/main/ui/Chat/ChannelSelect.tsx
@@ -18,7 +18,6 @@ export default function ChannelSelect({
       className="flex items-center bg-white rounded-l-xl pl-2 pr-0 h-full cursor-pointer select-none relative"
       onClick={onToggle}
     >
-      {/* 현재 채널 */}
       <span
         className={`text-[30px] font-bold pr-2 ${
           selected.value === '친구' ? 'text-[#305946]' : 'text-gray-800'
@@ -27,7 +26,6 @@ export default function ChannelSelect({
         {selected.label}
       </span>
 
-      {/* 화살표 */}
       <span
         className={`w-10 h-full flex items-center justify-center text-xl border-l border-r border-[#305946] ${
           isOpen ? 'bg-white text-[#305946]' : 'bg-[#305946] text-white'
@@ -36,10 +34,8 @@ export default function ChannelSelect({
         {isOpen ? '▲' : '▼'}
       </span>
 
-      {/* 드롭다운 */}
       {isOpen && (
         <div className="absolute left-0 top-[50px] w-[140px] bg-white border border-black rounded-xl shadow-lg z-50 overflow-hidden">
-          {/* 전체 */}
           <div
             onClick={(e) => {
               e.stopPropagation();
@@ -50,7 +46,6 @@ export default function ChannelSelect({
             전체
           </div>
 
-          {/* 친구 */}
           <div
             onClick={(e) => {
               e.stopPropagation();
