+++ src/features/main/ui/Chat/MentionDropdown.tsx
