+++ src/features/main/ui/Rooms/RoomCard.tsx
