+++ src/features/main/ui/Rooms/SearchBar.tsx
